Single and double snapping detection are working
The timeout is set to 1 s and the debouncing time is set to 100 ms.
Even with these values a single snapping is generating two
interruptions (consistently). Therefore, a simple solution was to
identify a single snapping with counting 2 interruptions and a
double snapping with 4 interruptions. We need to check this strange
behavior using a logic analyzer.