Sketch (pretty near of final version actually) of the FSM.
The code as it is right now logs the events that are occurring into
the console (UART0). The next step, instead of logging, is to
publish the correspoding  message on the corresponding topic
(pretty straightforward to adapt).

We still haven't the POSITIVE_ACCELERATION and
NEGATIVE_ACCELERATION events' detection, but all of the other
events are being properly detected.